+++ packages/fluent-ui/src/SubmitButton/SubmitButton.tsx
@@ -1,10 +1,16 @@
 import React from "react";
+import { utils, WidgetProps } from '@rjsf/core';
 import { PrimaryButton } from "@fluentui/react";
-export default () => (
-  <div>
-    <br />
-    <div className="ms-Grid-col ms-sm12">
-      <PrimaryButton text="Submit" type="submit" />
+const { getSubmitButtonProps } = utils;
+export default ({uiSchema}: WidgetProps) => {
+  const { submitText, required, ...submitButtonProps }= getSubmitButtonProps(uiSchema);
+  if(!required) return null;
+  return (
+    <div>
+      <br />
+      <div className="ms-Grid-col ms-sm12">
+        <PrimaryButton text={submitText} type="submit" {...submitButtonProps} />
+      </div>
     </div>
-  </div>
-);
+  );
+};