+++ packages/material-ui/src/SubmitButton/SubmitButton.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import Box from "@material-ui/core/Box";
+import { WidgetProps, utils } from '@rjsf/core';
+
+import Button from '@material-ui/core/Button';
+const { getSubmitButtonProps } = utils;
+const SubmitButton: React.FC<WidgetProps> = props => {
+  const { submitText, required, ...submitButtonProps }= getSubmitButtonProps(props.uiSchema);
+  if(!required) return null;
+  return (
+    <Box marginTop={3}>
+      <Button type="submit" variant="contained" color="primary" {...submitButtonProps}>
+        {submitText}
+      </Button>
+    </Box>
+  );
+};
+
+export default SubmitButton;