+++ packages/material-ui/src/Theme/Theme.tsx
@@ -1,8 +1,3 @@
-import React from "react";
-
-import Box from "@material-ui/core/Box";
-import Button from "@material-ui/core/Button";
-
 import ArrayFieldTemplate from "../ArrayFieldTemplate";
 import ErrorList from "../ErrorList";
 import Fields from "../Fields";
@@ -16,16 +11,7 @@ const { getDefaultRegistry } = utils;
 
 const { fields, widgets } = getDefaultRegistry();
 
-const DefaultChildren = () => (
-  <Box marginTop={3}>
-    <Button type="submit" variant="contained" color="primary">
-      Submit
-    </Button>
-  </Box>
-);
-
 const Theme: ThemeProps = {
-  children: <DefaultChildren />,
   ArrayFieldTemplate,
   fields: { ...fields, ...Fields },
   FieldTemplate,